pushExtensions param fix
I've altered the params passed to pushExtensions when it's called inside WatchExtensions to mirror the way it's called when `spicetify apply` is executed.

I'm new to Go, but since `watch -e` does not properly update extensions since the last release, and given that `spicetify apply` does update them as expected, this change seemed logical to me.