Open File explorer with -c and -config commands + Bug Fix
With mac, the config file was changed to a hidden folder for cleanness purposes, however users are typically unable to access those without following special steps (Which many unexperienced users may not be capable of.) This provides them with support. Also, before this commit -c and -config were printing the incorrect file path, this is now resolved.